requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==5.2.12
djangorestframework==3.17.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
